Added toString() to a bunch of Setting functions Added System.out.println "test" for SettingsResourceBundle Updated SettingsResourceBundle to use new Settings implementation

+++ community/cypher/src/main/scala/org/neo4j/cypher/ExecutionEngine.scala
@@ -93,7 +93,7 @@ class ExecutionEngine(graph: GraphDatabaseService, logger: StringLogger = String
   private def getQueryCacheSize() : Int = if (graph.isInstanceOf[InternalAbstractGraphDatabase]) {
     val database = graph.asInstanceOf[InternalAbstractGraphDatabase]
     database.getConfig.get(GraphDatabaseSettings.query_cache_size) match {
-      case v:java.lang.Integer => v
+      case v:java.lang.Long => v.intValue()
       case _ => 100
     }
   } else {
